+++ R/setup_ncit.R
@@ -1,29 +1,25 @@
 #' @title
-#' Run NCI Setup
+#' Run NCI Thesaurus Setup
 #'
 #' @description
-#' Load the NCI Metathesaurus RRF files, followed by the
-#' NCI Thesaurus and the Code to CUI Map.
+#' Load the NCI Thesaurus and the Code to CUI Map.
 #'
 #' @inheritParams package_arguments
 #' @inheritParams load_ncim
 #'
-#' @rdname run_setup
+#' @rdname setup_ncit
 #'
 #' @export
 
-run_setup <-
+setup_ncit <-
   function(conn,
            conn_fun = "pg13::local_connect()",
-           schema = "nci",
+           schema = "ncit",
            path_to_rrfs,
            ncim_version,
            log_schema = "public",
-           log_table = "setup_nci_log") {
+           log_table = "setup_ncit_log") {
 
-    .Deprecated(new = "setup_ncim or setup_ncit")
-
-    secretary::press_enter()
 
     if (missing(conn)) {
 
@@ -34,14 +30,6 @@ run_setup <-
 
 
     }
-
-    load_ncim(conn = conn,
-              schema = schema,
-              path_to_rrfs = path_to_rrfs,
-              log_schema = log_schema,
-              log_table = log_table,
-              ncim_version = ncim_version)
-
 
     load_ncit(conn = conn,
               schema = schema,